fix(e2e): upgrade to node 24 (GH action only), refactor utils to avoid copypasted functions
